Move mocks to a seperate file
